[update]URL変更によるパスの影響修正
