+++ src/app/api/network/getMatchGames/route.ts
@@ -1,4 +1,3 @@
-import { gameDetailType } from "@/types/api/gameDetailsType";
 import { steamGameCategoryType } from "@/types/api/steamDataType";
 import { NextResponse } from "next/server";
 
@@ -7,38 +6,44 @@ export async function GET() {
   try {
 
     const res = await fetch(`${process.env.NEXT_PUBLIC_CURRENT_URL}/api/getCurrentTopGames`);
-    const data = await res.json();
-  
-    const twitchGameId = data[0].twitchGameId;
-    const steamGameId = data[0].steamGameId;
+    const data = await res.json();  
 
-    const response = await fetch(`https://store.steampowered.com/api/appdetails?appids=${steamGameId}&cc=jp`);
-    const responseJson = await response.json();
-    const gameDetailData = responseJson[steamGameId]?.data;
+    const result = [];
+    for(let i = 0; i < data.length; i += 1) {
 
-    if (!gameDetailData) {
-      throw new Error(`Game details not found for Steam Game ID: ${steamGameId}`);
-    }
+      const twitchGameId = data[i].twitch_id;
+      const steamGameId = data[i].steam_id;
 
-    const result: gameDetailType = {
-      twitchGameId: twitchGameId,
-      steamGameId: steamGameId,
-      title: gameDetailData.name,
-      imgURL: gameDetailData.header_image,
-      gameData: {
-        genres: gameDetailData.genres,
-        price: gameDetailData.price_overview ? gameDetailData.price_overview.final : 0,
-        isSinglePlayer: gameDetailData.categories.some((category: steamGameCategoryType) => category.id === 2),
-        isMultiPlayer: gameDetailData.categories.some((category: steamGameCategoryType) => category.id === 1),
-        platforms: {
-          windows: gameDetailData.platforms.windows,
-          mac: gameDetailData.platforms.mac,
-          linux: gameDetailData.platforms.linux
-        }
+
+      const response = await fetch(`https://store.steampowered.com/api/appdetails?appids=${steamGameId}&cc=jp`);
+      const responseJson = await response.json();
+      if (!responseJson[steamGameId].success) {
+        console.error('Failed to fetch game detail data:', steamGameId);
+        continue;
       }
-    };
+      const gameDetailData = responseJson[steamGameId].data
 
-    return NextResponse.json(data);
+      result.push({
+        twitchGameId: twitchGameId,
+        steamGameId: steamGameId,
+        title: gameDetailData.name,
+        imgURL: gameDetailData.header_image,
+        gameData: {
+          genres: gameDetailData.genres,
+          price: gameDetailData.price_overview ? gameDetailData.price_overview.final : 0,
+          isSinglePlayer: gameDetailData.categories.some((category: steamGameCategoryType) => category.id === 2),
+          isMultiPlayer: gameDetailData.categories.some((category: steamGameCategoryType) => category.id === 1),
+          platforms: {
+            windows: gameDetailData.platforms.windows,
+            mac: gameDetailData.platforms.mac,
+            linux: gameDetailData.platforms.linux
+          }
+        }
+      })
+    }
+    
+
+    return NextResponse.json(result);
 
   } catch (error) {
     console.error("Error fetching top games:", error);
