+++ src/app/desktop/details/[steamGameId]/[twitchGameId]/page.tsx
@@ -18,7 +18,6 @@ export default function Page({
         <SimilarGames />
       </div>
       <div className="w-3/4 bg-[#2a475e] flex flex-col p-4">
-        <Breadcrumb steamGameId={steamGameId} twitchGameId={twitchGameId}/>
         <div className="basis-1/10">
           <GameTitle steamGameId={steamGameId} />
         </div>
@@ -37,24 +36,3 @@ export default function Page({
     </div>
   )
 }
-
-type BreadcrumbProps = {
-  steamGameId: string;
-  twitchGameId: string;
-};
-
-const Breadcrumb: React.FC<BreadcrumbProps> = ({ steamGameId, twitchGameId }) => {
-  return (
-    <div className="text-gray-300">
-      {/* <Link href="/network" legacyBehavior>
-        <a className="text-white text-gray-300 hover:text-gray-400">
-          ネットワーク
-        </a>
-      </Link>
-      ＞
-      <a className="text-white text-gray-300 hover:text-gray-400">
-        {steamGameId}/{twitchGameId}
-      </a> */}
-    </div>
-  );
-};
